Add the ability to auto import handlers
Having to explicitly import modules so that their handlers get
registered on the registry is tedious. You have to make sure you import
any new file.

Generally, these handlers are organized by directory for each registry,
if that is the case, we should just import all of the modules in that
directory to have the handlers be registered.

I also redid the todo app to use this method.